# API/api-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Video Game Exchange API",
    "description": "An API for listing and trading Retro Video Games between users"
  },
  "servers": [
    {
      "url": "http://localhost:5000/",
      "description": "Locally hosted node server port"
    }
  ],
  "tags": [
    {
      "name": "",
      "description": ""
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "description": "Get a list of Users registered in the Exchange",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "Users Retrieved Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type" : "array",
                  "items" : {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "links": {
              "GetUserByUserId": {
                "operationId": "getUser",
                "parameters": {
                  "id": "$id"
                },
                "description": "The `id` value returned in each of the users in the response can be used as the `id` parameter in `GET /users/{id}`."
              }
            }
          },
          "404": {
            "description": "No Users Found",
            "links": {
              "AddUser": {
                "operationId": "addUser",
                "description": "A User can be registered to the Exchange using `POST /users`."
              }
            }
          }
        }
      },
      "post": {
        "description": "Register a new User to the Exchange",
        "operationId": "addUser",
        "responses": {
          "201": {
            "description": "User Created Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "links" : {
              "GetUserByUserId": {
                "operationId": "getUser",
                "parameters": {
                  "id": "$id"
                },
                "description": "The `id` value returned in the response can be used as the `id` parameter in `GET /users/{id}`."
              }
            }
          },
          "400": {
            "description": "Missing Required Fields"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Returns a User by id",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User Retrieved Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "links" : {
              "GetUserGames": {
                "operationId": "getUserGames",
                "parameters": {
                  "id": "$id"
                },
                "description": "The `id` value returned in the response can be used as the `id` parameter in `GET /users/{id}/games`."
              }
            }
          },
          "400": {
            "description": "Invalid User Id"
          },
          "404": {
            "description": "User Not Found"
          }
        }
      },
      "patch": {
        "description": "Update allowed fields on a User",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User Update Successful",
            "links" : {
              "GetUserByUserId": {
                "operationId": "getUser",
                "parameters": {
                  "id": "$id"
                },
                "description": "The `id` value returned in the response can be used as the `id` parameter in `GET /users/{id}`."
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "User Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      },
      "delete": {
        "description": "Delete a User by id",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User Deleted Successfully",
            "links": {
              "AddUser": {
                "operationId": "addUser",
                "description": "A User can be registered to the Exchange using `POST /users`."
              }
            }
          },
          "400": {
            "description": "Invalid User Id"
          },
          "404": {
            "description": "User Not Found"
          }
        }
      }
    },
    "/users/{id}/games": {
      "get": {
        "description": "Retrive the list of Games owned by a User",
        "operationId": "getUserGames",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Games List Retrieved Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type" : "array",
                  "items" : {
                    "$ref": "#/components/schemas/Game"
                  }
                }
              }
            },
            "links": {
              "GetGameByGameId": {
                "operationId": "getGame",
                "parameters": {
                  "id": "$id"
                },
                "description": "The `id` value returned in each of the games in the response can be used as the `id` parameter in `GET /games/{id}`."
              }
            }
          },
          "400": {
            "description": "Invalid User Id"
          },
          "404": {
            "description": "User Not Found"
          }
        }
      }
    },
    "/games": {
      "get": {
        "description": "Get Games listed on the Exchange",
        "operationId": "getGames",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "publisher",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "system",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "condition",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Games Retrieved Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type" : "array",
                  "items" : {
                    "$ref": "#/components/schemas/Game"
                  }
                }
              }
            },
            "links": {
              "GetGameByGameId": {
                "operationId": "getGame",
                "parameters": {
                  "id": "$id"
                },
                "description": "The `id` value returned in each of the games in the response can be used as the `id` parameter in `GET /games/{id}`."
              }
            }
          }
        }
      },
      "post": {
        "description": "Add a Game to the Exchange",
        "operationId": "addGame",
        "responses": {
          "201": {
            "description": "Game Created Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            },
            "links": {
              "GetGameByGameId": {
                "operationId": "getGame",
                "parameters": {
                  "id": "$id"
                },
                "description": "The `id` value returned in the response can be used as the `id` parameter in `GET /games/{id}`."
              }
            }
          },
          "400": {
            "description": "Missing Required Fields"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Game"
              }
            }
          }
        }
      }
    },
    "/games/{id}": {
      "get": {
        "description": "Returns a Game by id ",
        "operationId": "getGame",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game Retrieved Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            },
            "links": {
              "GetUserByUserId": {
                "operationId": "getUser",
                "parameters": {
                  "id": "$id"
                },
                "description": "The `ownerId` value returned in the response can be used as the `id` parameter in `GET /users/{id}`."
              },
              "GetGamesByUser": {
                "operationId": "getUserGames",
                "parameters": {
                  "id": "$id"
                },
                "description": "The `ownerId` value returned in the response can be used as the `id` parameter in `GET /users/{id}/games`."
              }
            }
          },
          "400": {
            "description": "Invalid Game Id"
          },
          "404": {
            "description": "Game Not Found"
          }
        }
      },
      "put": {
        "description": "Updates all the fields of a Game",
        "operationId": "fullUpdateGame",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full Game Update Successful",
            "links": {
              "GetGameByGameId": {
                "operationId": "getGame",
                "parameters": {
                  "id": "$id"
                },
                "description": "The `id` value returned in the response can be used as the `id` parameter in `GET /games/{id}`."
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Game Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Game"
              }
            }
          }
        }
      },
      "patch": {
        "description": "Updates specified fields of a Game",
        "operationId": "partialUpdateGame",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Partial Game Update Successful",
            "links": {
              "GetGameByGameId": {
                "operationId": "getGame",
                "parameters": {
                  "id": "$id"
                },
                "description": "The `id` value returned in the response can be used as the `id` parameter in `GET /games/{id}`."
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Game Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Game"
              }
            }
          }
        }
      },
      "delete": {
        "description": "Delete a Game by id",
        "operationId": "deleteGame",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game Deleted Successfully",
            "links": {
              "AddGame": {
                "operationId": "addGame",
                "description": "A Game can be added to the Exchange using `POST /games`."
              }
            }
          },
          "400": {
            "description": "Invalid Game Id"
          },
          "404": {
            "description": "Game Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Game" : {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 0
          },
          "name": {
            "type": "string",
            "example": "Mario 64"
          },
          "publisher": {
            "type": "string",
            "example": "Nintendo"
          },
          "year": {
            "type": "integer",
            "example": 1996
          },
          "system": {
            "type": "string",
            "example": "Nintendo 64"
          },
          "condition": {
            "type": "string",
            "example": "Mint"
          },
          "ownerId": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 0
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@mail.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "address": {
            "type": "string",
            "example": "1 Main Street City, State"
          }
        }
      }
    }
  }
}